+++ Release-Notes.md
@@ -3,6 +3,7 @@
 ## Unreleased Fixes
 
 - Get-PSLocalGallery - Fixed accidental method call on a variable with zero objects
+- New Command - Register-PSLocalGallery - Registers the PSLocalGallery PSRepository
 
 ## 1.0.7